% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{auto_color}
\alias{auto_color}
\title{Automatically choose color scale for categorical data.}
\usage{
auto_color(...)
}
\arguments{
\item{...}{common discrete scale parameters: \code{name}, \code{breaks},
\code{labels}, \code{na.value}, \code{limits} and \code{guide}.  See
\code{\link{discrete_scale}} for more details}
}
\description{
Uses color brewer qualitative Set1 (without the yellow) for data with up to
8 levels and the default ggplot2 color scale otherwise.
}
\examples{
d <- diamonds[sample(nrow(diamonds), 5000), ]
p <- ggplot(d, aes(carat, price))

# fewer than 9 levels
p + geom_point(aes(color = cut(price, breaks = 4))) + auto_color()
p + geom_point(aes(color = cut(price, breaks = 8))) + auto_color()

# more than 8 levels
p + geom_point(aes(color = cut(price, breaks = 9))) + auto_color()
p + geom_point(aes(color = cut(price, breaks = 50))) + auto_color(guide = FALSE)
}

