% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_pres}
\alias{theme_pres}
\title{Theme for slide presentations}
\usage{
theme_pres()
}
\description{
Provides sensible defaults for graphs intended for projection
display.
}
\examples{
p <- ggplot(mtcars, aes(mpg, wt)) + geom_point()
p
p + theme_pres()
}

