% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_pub}
\alias{theme_pub}
\title{Publication-ready theme.}
\usage{
theme_pub(trim_white_space = FALSE)
}
\description{
Provides sensible defaults for graphs intended for publication, including
using a white background, setting appropriate font sizes, removing white
space, and adjusting line widths.
}
\examples{
p <- ggplot(mtcars, aes(mpg, wt)) + geom_point()
p
p + theme_pub()
}

