% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{tilt_x_labels}
\alias{tilt_x_labels}
\title{Tilt x-axis labels.}
\usage{
tilt_x_labels(angle = 45, hjust = 1, vjust = 1)
}
\arguments{
\item{angle}{The angle text should be displayed at (0 is horizontal, 90 is
vertical)}

\item{hjust}{horizontal justification (0 to 1)}

\item{vjust}{vertical justification (0 to 1)}
}
\description{
Tilt x-axis labels.
}
\examples{
mtcars$make <- sapply(rownames(mtcars), function(x) strsplit(x, " ")[[1]][1])
p <- ggplot(mtcars, aes(make, wt)) + geom_point()
p
p + tilt_x_labels()
}

